+++ common/src/main/java/software/amazon/s3/dataaccelerator/request/Range.java
@@ -29,7 +29,7 @@ public class Range {
   long end;
 
   private static final String TO_HTTP_STRING_FORMAT = "bytes=%d-%d";
-  private static final String TO_STRING_FORMAT = "[%d-%d]";
+  private static final String TO_STRING_FORMAT = "%d-%d";
 
   /**
    * Construct a range. At least one of the start and end of range should be present.